Create progressBar
This is the VBA code I use to add a dynamic progress bar to my slideshows. 

This function takes the number of slides in your slideshow, and 'fills in' progress bars depending on where you are in your slideshow. You can re-run the Macro to modify the previously-created progress bar as well (repeated use). 

******* VARIABLES TO MODIFY ********
barColor --> RGB tripplet representing your target line and fill color
gapRatio --> the side of your gaps, relative to the 'bars' 
dashHeight --> the height of the bars
bottomOffset --> how far from the bottom of the slide you want the bar to be (in points)

****** CAVEAT ******
This sub does not detect a title page. If you only want the progress bar to start on Slide 2, you'll have to temporarily remove the title page, run the macro, and then add the title page back in.